network, common: move VIF struct to cache.types
